·fix: fixed update method.

+++ JwtAuthenticationManager/Domain/Services/IUserService.cs
@@ -6,9 +6,9 @@ namespace JwtAuthenticationManager.Domain.Services;
 public interface IUserService
 {
     Task<IEnumerable<User>> ListAsync();
-    Task<User> GetByIdAsync(int id);
+    Task<User> GetByIdAsync(long id);
     Task<AuthenticationResponse?> Authenticate(AuthenticationRequest authenticationRequest);
     Task RegisterAsync(RegisterRequest request);
-    Task UpdateAsync(int id, UpdateRequest request);
-    Task DeleteAsync(int id);
+    Task UpdateAsync(long id, UpdateRequest request);
+    Task DeleteAsync(long id);
 }
